i18n + cleanup: complete chrome translations; fix compiler warnings
Translations (FR/DE):
- Wrap the remaining untranslated chrome in TRANS(): group panel titles
  (decouple the stable English name used for decoration matching from the
  displayed text), context menu (Reset to default / Randomize), Multi page
  (Omni, "Editing Part X of Y", "Part N"), file dialogs, PresetBrowser
  (Templates / Multi).
- Add ~35 keys to both FR + DE tables; fill gaps for already-TRANS'd keys
  (Font, Console, Serif, Sans Serif, Length, Ambika/Parvati Patch).
- ParamPage::refreshLanguage() re-applies group titles live on language switch.

Warnings (clean build, 2 intentional exceptions):
- SynthEngine.h/.cpp: int->size_t index casts; fix voices[i] (OwnedArray is
  int-indexed, the (size_t) cast was wrong); std::any_of; redundant-cast removal.
- PluginEditor.cpp: sign-conversion casts, shadow renames, const-correctness,
  De Morgan, std::any_of; drop unused PatchFile.h include.
- AmbikaVoice / PatchFile / PluginProcessor / Arpeggiator / ParvatiLookAndFeel /
  IconButton / ParameterLayout: int->float + int->size_t casts, negation cast,
  remove deprecated C++17 redundant constexpr static def, drop unused <algorithm>.
- Left deliberately: -Wfloat-equal on the dsp/analog_filter cutoff/resonance
  cache guards (bit-identical-audio constraint).

Verified: polyphony/multitimbral/arp/sequencer/roundtrip/editor tests pass;
the 2 patch_load_test failures are pre-existing on HEAD (unrelated).

+++ Source/AmbikaVoice.cpp
@@ -2,7 +2,6 @@
 
 #include "AmbikaVoice.h"
 
-#include <algorithm>
 #include <cmath>
 
 #include "AmbikaSound.h"
@@ -171,7 +170,7 @@ void AmbikaVoice::startNote (int midiNoteNumber, float velocity,
     const int note14  = juce::jlimit (0, static_cast<int> (ambika::dsp::kHighestNote),
                                      baseNote * 128 + partTuning_ + spreadDrift14_);
     const int velInt = juce::jlimit (0, 255, static_cast<int> (velocity * 255.0f));
-    voice_.Trigger (note14, static_cast<uint8_t> (velInt & 0xFF), legato ? 1 : 0);
+    voice_.Trigger (static_cast<uint16_t> (note14), static_cast<uint8_t> (velInt & 0xFF), legato ? 1 : 0);
     legatoNext_ = false;   // one-shot hint, consumed
     spreadDrift14_ = 0;    // one-shot hint, consumed (default trigger => no drift)
 
@@ -336,7 +335,7 @@ void AmbikaVoice::fillInternalBlock()
         for (int i = 0; i < ambika::dsp::kAudioBlockSize; ++i)
         {
             // 8-bit (centred 128) -> float. Keep the (128 vs 127) asymmetry exact.
-            float s = (static_cast<int> (out[i]) - 128) / 128.0f;
+            float s = static_cast<float> (static_cast<int> (out[i]) - 128) / 128.0f;
             s = filter_.processSample (s);
             s *= vcaGain;
             fifo_.push_back (s);
@@ -383,7 +382,7 @@ void AmbikaVoice::fillInternalBlock()
     for (int i = 0; i < ambika::dsp::kAudioBlockSize; ++i)
     {
         // 8-bit (centred 128) -> float. Keep the (128 vs 127) asymmetry exact.
-        float s = (static_cast<int> (out[i]) - 128) / 128.0f;
+        float s = static_cast<float> (static_cast<int> (out[i]) - 128) / 128.0f;
         filter_.setCutoffHz (smoothedCutoffHz_.getNextValue());
         filter_.setResonance (smoothedResonance_.getNextValue());
         filter_.commit();
@@ -402,7 +401,7 @@ void AmbikaVoice::fillInternalBlock()
         // linear so no aliasing) -> FIFO.
         float raw[ambika::dsp::kAudioBlockSize];
         for (int i = 0; i < ambika::dsp::kAudioBlockSize; ++i)
-            raw[i] = (static_cast<int> (out[i]) - 128) / 128.0f;
+            raw[i] = static_cast<float> (static_cast<int> (out[i]) - 128) / 128.0f;
 
         // Upsample 40 -> 40*N into the Oversampling internal buffer.
         const float* inChannels[1] = { raw };
